backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
websockets>=11.0.3
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
